Magic variable implementation
This allows you to do Packer::Config#variable.<anything> to get a packer.io variable reference string. It's definitely cleaner syntax in the template spec but it has some issues. I'm going to sit on it for a bit before I add it to Packer::Config. This commit just implements Packer::Variable and its tests to show the idea.